fix(sdk): resolve Dockerfile launch annotations
Move DockerfileLaunch into a lightweight value module so public annotations resolve at runtime without loading the parser or a backend. Keep parser and backend imports lazy, and calibrate the supported-capability lifecycle wording to allow API evolution with migration guidance.
